cliinterface: Wait a few seconds before killing the process.
In some cases, failOperation() will try to kill an application which
is already in the process of finishing gracefully, so it is better to
wait a few seconds before trying to kill the process.

svn path=/trunk/KDE/kdeutils/ark/; revision=1233140

+++ kerfuffle/cliinterface.cpp
@@ -654,7 +654,10 @@ bool CliInterface::checkForErrorMessage(const QString& line, int parameterIndex)
 bool CliInterface::doKill()
 {
     if (m_process) {
-        m_process->kill();
+        // Give some time for the application to finish gracefully
+        if (!m_process->waitForFinished(5)) {
+            m_process->kill();
+        }
 
         return true;
     }
